requirements.txt
boto3==1.28.38
python-dotenv==1.0.0
scikit-learn==1.6.1
requests 